⚡ Bolt: Elide bounds checks in Sobel filter (#1079)
- Extract explicit 1D slices for `prev_row`, `curr_row`, and `next_row` outside the inner pixel loop in `apply_sobel` (and its SIMD fallback).
- By providing the compiler with the explicit `width` bound, it can mathematically prove safety and safely elide 9 per-pixel bounds checks.
- Addressed various `clippy` and formatting issues via `cargo clippy --fix` and `cargo fmt`.

+++ examples/melt_demo.rs
@@ -54,7 +54,7 @@ impl MeltDemoApp {
         // Title text approximation (just a block for effect)
         for y in 50..100 {
             for x in 50..400 {
-                background_fb.set_pixel(x as i32, y as i32, 0xFF_DD_22_22); // Red block
+                background_fb.set_pixel(x, y, 0xFF_DD_22_22); // Red block
             }
         }
